+++ Sharpliner.SourceGenerator/StringConditionGenerator.cs
@@ -1,6 +1,4 @@
-﻿
-using System.Collections.Immutable;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Text;
 using Microsoft.CodeAnalysis;
