+++ src/Sharpliner/SharplinerSerializer.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Text.RegularExpressions;
-using YamlDotNet.Core;
 using YamlDotNet.Serialization;
-using YamlDotNet.Serialization.EventEmitters;
 using YamlDotNet.Serialization.NamingConventions;
 
 namespace Sharpliner;
